requirements.txt
ultralytics
streamlit
pillow
numpy
pandas
